+++ tests/unit/test_api.py
@@ -116,6 +116,7 @@ WEBVIEW_PROPERTIES = (
     "back_forward_gestures",
     "default_context_menus",
     "https_scheme",
+    "proxy",
     "navigation_allow",
     "open_external",
     "download_allow",